feat(health-check): replace the database based health check by a open-telemetry based health check system
gh/feature/19
