--- adapters/diesel/src/repositories/health_check_info/mod.rs
@@ -1,3 +1,0 @@
-mod health_check_info_write;
-
-pub(crate) use health_check_info_write::*;
