+++ core/src/models/mod.rs
@@ -2,6 +2,6 @@ mod account_life_cycle_config;
 mod config;
 mod webhook_config;
 
-pub use account_life_cycle_config::AccountLifeCycle;
-pub use config::CoreConfig;
-pub use webhook_config::WebhookConfig;
+pub use account_life_cycle_config::*;
+pub use config::*;
+pub use webhook_config::*;
