+++ http_tools/src/models/mod.rs
@@ -1,3 +1,4 @@
+pub mod api_otel_codes;
 pub mod auth_config;
 pub mod external_providers_config;
 pub mod internal_auth_config;
