+++ settings/config.for-docker.yaml
@@ -163,7 +163,7 @@ api:
   # ? --------------------------------------------------------------------------
   # ? Health check settings
   # ? --------------------------------------------------------------------------
-  healthCheckInterval: 30
+  healthCheckInterval: 10
   maxRetryCount: 3
 
   # ? --------------------------------------------------------------------------
